apps/backend/requirements.txt
# apps/backend/requirements.txt  (Python 3.11.x)

# Web framework + server (local dev)
fastapi==0.119.0
uvicorn[standard]==0.37.0

# Lambda adapter
mangum==0.19.0

# Validation / models
pydantic==2.12.1
pydantic-settings==2.5.2

# Mongo async stack (Beanie ODM + Motor)
beanie==2.0.0
motor==3.7.1

# Redis (EV, permset cache, JTI blocklist)
redis[hiredis]==5.0.8

# Auth / crypto
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client (e.g., Supabase JWKS / admin calls)
httpx==0.27.2

# Forms / multipart (if needed by any route)
python-multipart==0.0.9

# Nice-to-haves used in examples/middleware
itsdangerous==2.2.0
loguru==0.7.2

# ---- Dev / test (optional) ----
pytest==8.3.3
pytest-asyncio==0.23.7


PyJWT>=2.8.0
